Moved base elasticsearch logic to common package.
