--- src/main/java/uk/nhs/ctp/auditFinder/role/RoleMapper.java
@@ -1,7 +1,0 @@
-package uk.nhs.ctp.auditFinder.role;
-
-public interface RoleMapper {
-
-  void setupSupplierRoles(String supplierId, String username);
-
-}
