+++ src/main/java/uk/nhs/ctp/service/UserManagementService.java
@@ -9,9 +9,9 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import uk.nhs.cactus.common.elasticsearch.role.RoleMapper;
 import uk.nhs.cactus.common.security.JWTHandler;
 import uk.nhs.cactus.common.security.JWTRequest;
-import uk.nhs.ctp.auditFinder.role.RoleMapper;
 import uk.nhs.ctp.entities.UserEntity;
 import uk.nhs.ctp.exception.EMSException;
 import uk.nhs.ctp.model.RegisterSupplierRequest;
