+++ src/test/java/uk/nhs/ctp/service/UserManagementServiceTest.java
@@ -23,9 +23,9 @@ import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.test.util.ReflectionTestUtils;
+import uk.nhs.cactus.common.elasticsearch.role.RoleMapper;
 import uk.nhs.cactus.common.security.JWTHandler;
 import uk.nhs.cactus.common.security.JWTRequest;
-import uk.nhs.ctp.auditFinder.role.RoleMapper;
 import uk.nhs.ctp.entities.UserEntity;
 import uk.nhs.ctp.exception.EMSException;
 import uk.nhs.ctp.model.RegisterSupplierRequest;
